requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
pandas
numpy
scikit-learn
google-genai
